+++ src/components/Works/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Icon1 from '../../assets/images/Cricket_Logo.png';
 import Icon2 from '../../assets/images/Lacrosse_works.png';
-import Icon3 from '../../assets/images/MovieApp.png';
+import Icon3 from '../../assets/images/ysfilmImg.png';
 import Icon4 from '../../assets/images/workDescFour.png';
 import Icon5 from '../../assets/images/covidDash_logo.png';
 import Icon6 from '../../assets/images/portfolioImg.png';
